+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Master (unreleased)
 
 ## Enhancements
 * Allow early running exit tasks and avoid the `at_exit` hook through the `SimpleCov.run_exit_tasks!` method. (thanks [@macumber]: https://github.com/macumber))
+* Allow manual collation of result sets through the `SimpleCov.collate` entrypoint. See the README for more details (thanks [@ticky](https://github.com/ticky))
 
 0.18.0.beta2 (2020-01-05)
 ===================